test(storagenode): do not delete ".keep" file from `internal/storagenode/testdata/relative_volume`
This patch makes test not delete `internal/storagenode/testdata/relative_volume/.keep`.
Hidden file `internal/storagenode/testdata/relative_volume/.keep` prevents
`internal/storagenode/testdata/relative_volume` from deleting since git is ignore empty directory.

+++ internal/storagenode/storagenode_test.go
@@ -419,6 +419,9 @@ func TestStorageNode_MakeVolumesAbsolute(t *testing.T) {
 		ps, err := filepath.Glob("./testdata/relative_volume/*")
 		assert.NoError(t, err)
 		for _, p := range ps {
+			if filepath.Base(p) == ".keep" {
+				continue
+			}
 			_ = os.RemoveAll(p)
 		}
 	}()
